+++ NEWS.md
@@ -1,3 +1,13 @@
+# avocado 0.2.0
+
+## Breaking Changes
+* Column names changed to be more descriptive
+
+## Improvements
+* Datasets updated to include data through 2024 calendar year
+* Units definition clarified to mean avocados instead of weight
+* Reduced vignettes down to only 1 vignette for clarity
+
 # avocado 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.